# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/office-hours": {
      "get": {
        "operationId": "OfficeHoursController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "teacherCoreUserId",
            "required": false,
            "in": "query",
            "description": "ค่า sub ของอาจารย์ (ไม่ใช่ UUID เสมอไป)",
            "schema": {
              "maxLength": 64,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/OfficeHourResponse"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMetaModel"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "เวลาทำการรายสัปดาห์ของอาจารย์ (กรองด้วย teacherCoreUserId)",
        "tags": [
          "office-hours"
        ]
      },
      "post": {
        "operationId": "OfficeHoursController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOfficeHourDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/OfficeHourResponse"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "เพิ่มเวลาทำการของตัวเอง",
        "tags": [
          "office-hours"
        ]
      }
    },
    "/api/v1/office-hours/{id}": {
      "patch": {
        "operationId": "OfficeHoursController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOfficeHourDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/OfficeHourResponse"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "แก้ไขเวลาทำการ",
        "tags": [
          "office-hours"
        ]
      },
      "delete": {
        "operationId": "OfficeHoursController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeletedModel"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ลบเวลาทำการ",
        "tags": [
          "office-hours"
        ]
      }
    },
...
    "/api/v1/chat-messages": {
      "get": {
        "operationId": "ChatMessagesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "bookingId",
            "required": true,
            "in": "query",
            "description": "นัดหมายที่ต้องการอ่านข้อความ",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ChatMessageResponse"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMetaModel"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ข้อความในนัดหมายหนึ่งรายการ (เก่า → ใหม่)",
        "tags": [
          "chat-messages"
        ]
      },
      "post": {
        "operationId": "ChatMessagesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChatMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ChatMessageResponse"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ส่งข้อความในนัดหมายของตัวเอง",
        "tags": [
          "chat-messages"
        ]
      }
    },
    "/api/v1/chat-messages/{id}/read": {
      "patch": {
        "operationId": "ChatMessagesController_markRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ChatMessageResponse"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ทำเครื่องหมายว่าอ่านข้อความแล้ว",
        "tags": [
          "chat-messages"
        ]
      }
    },
...
  },
  "info": {
    "title": "Q-Teacher API",
    "description": "ระบบนัดหมายเข้าพบอาจารย์ (CSMJU2030) — ตอบเป็น { success, data[, meta] } หรือ { success: false, error } ตามมาตรฐาน api-conventions.md · ต้องมี token ของ Core Hub ทุก endpoint ยกเว้น /api/health",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "csmju_q_teacher_access_token"
      }
    },
    "schemas": {
      "PaginationMetaModel": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 42
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "totalPages": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
      "OfficeHourResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "teacherCoreUserId": {
            "type": "string",
            "description": "ค่า sub ของอาจารย์เจ้าของเวลา"
          },
          "teacherPersonCode": {
            "type": "string",
            "nullable": true,
            "description": "รหัสบุคลากรจาก Core Hub (แสดงแทนชื่อ)"
          },
          "dayOfWeek": {
            "type": "string",
            "enum": [
              "MON",
              "TUE",
              "WED",
              "THU",
              "FRI"
            ]
          },
          "startTime": {
            "type": "string",
            "example": "09:00",
            "description": "เวลาท้องถิ่น Asia/Bangkok รูปแบบ HH:mm"
          },
          "endTime": {
            "type": "string",
            "example": "12:00"
          },
          "slotMinutes": {
            "type": "number",
            "example": 30
          },
          "isAvailable": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "teacherCoreUserId",
          "teacherPersonCode",
          "dayOfWeek",
          "startTime",
          "endTime",
          "slotMinutes",
          "isAvailable",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateOfficeHourDto": {
        "type": "object",
        "properties": {
          "dayOfWeek": {
            "type": "string",
            "enum": [
              "MON",
              "TUE",
              "WED",
              "THU",
              "FRI"
            ],
            "example": "MON"
          },
          "startTime": {
            "type": "string",
            "example": "09:00",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$"
          },
          "endTime": {
            "type": "string",
            "example": "12:00",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$"
          },
          "slotMinutes": {
            "type": "number",
            "minimum": 5,
            "maximum": 240,
            "default": 30,
            "description": "ความยาวของช่วงนัด 1 ช่วง (นาที)"
          },
          "isAvailable": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "dayOfWeek",
          "startTime",
          "endTime"
        ]
      },
      "UpdateOfficeHourDto": {
        "type": "object",
        "properties": {
          "dayOfWeek": {
            "type": "string",
            "enum": [
              "MON",
              "TUE",
              "WED",
              "THU",
              "FRI"
            ],
            "example": "MON"
          },
          "startTime": {
            "type": "string",
            "example": "09:00",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$"
          },
          "endTime": {
            "type": "string",
            "example": "12:00",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$"
          },
          "slotMinutes": {
            "type": "number",
            "minimum": 5,
            "maximum": 240,
            "default": 30,
            "description": "ความยาวของช่วงนัด 1 ช่วง (นาที)"
          },
          "isAvailable": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "DeletedModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "deleted": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "id",
          "deleted"
        ]
      },
...
      "ChatMessageResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "bookingId": {
            "type": "string",
            "format": "uuid"
          },
          "senderCoreUserId": {
            "type": "string",
            "description": "ค่า sub ของผู้ส่ง — เทียบกับ id จาก GET /api/v1/me เพื่อรู้ว่าเป็นข้อความของเราหรือไม่"
          },
          "message": {
            "type": "string"
          },
          "isRead": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "bookingId",
          "senderCoreUserId",
          "message",
          "isRead",
          "createdAt"
        ]
      },
      "CreateChatMessageDto": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "string",
            "format": "uuid"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000,
            "example": "สวัสดี ขอเลื่อนนัดได้ไหม"
          }
        },
        "required": [
          "bookingId",
          "message"
        ]
      },
...
    }
  }
}